Nuxt starter #4 (#6)
* Initial app skeleton via nuxt-starter template

> https://github.com/nuxt/starter

* Use SCSS for all styles

Via sass-loader and node-sass.

* Layout and dynamic route for OKU circ

Demonstrates:
* Introducing styles at layout level

* Passing URL params to the page component via dynamic routing

* Copy template for page component from playground

To help better illustrate what we're about to do...

* Fetch Olin laptop data from libservices

Initially used middleware(1) for fetching, but abandoned that approach
in favor of modular Vuex stores(2). Both approaches rely on Axios via
nuxt-helpers(3).

Proof of concept, so beware of the following:

* Data is fetched only once -- prior to initial page load
  (aka no auto updating just yet)

* While dynamic routing for page can accept any `location`, the API call
  is not dynamic (or location aware) and is hardcoded for Olin right now

* nuxt-helpers will soon be replaced by official nuxt-modules(4)

* Location param is passed to laptops store via second payload argument
  -- ideally the location name would be returned by the API

* In this first pass, no logic to determine if NextAvailable returned
  from API is actually a future date

Bottomline: Getting very close to abandoning libservices and writing a
new microservice.

(1) https://nuxtjs.org/guide/routing#middleware
(2) https://nuxtjs.org/guide/vuex-store
(3) https://github.com/fandogh/nuxt-helpers
(4) https://github.com/nuxt/modules

* Uppercase the location param for rendering

Add lodash via yarn and include in vendor bundle.

* Fetch Olin phone charger data from libservices

All of the caveats from the earlier commit for laptops (42d0f7e) applies
here as well.

Since the equipment feed is even messier than laptops, more massaging is
required, which has been handed off to a new libservices util in an
attempt to keep the store mutations and actions clean.

* First pass at oku-circ component

Call it twice from the page, passing the location prop and this results
in rendering the status for both locations.

This includes updating the libservices util to support dynamic API calls
(both in terms of endpoints and location ID).

Couldn't set the API_URL env variable as eluded to in the nuxt-helpers
doc, so resorted to declaring a constant within the util. This is all
just proof-of-concept for now anyway...everything is subject to change.

https://github.com/fandogh/nuxt-helpers#-usage

It should be noted that the location passed in the dynamic route
`/oku/:location` has no consequence on the oku-circ component just yet;
Olin will always be rendered first. Similarily, the page's fetch method
will always dispatch the laptops and phoneChargers actions for both Olin
and Uris, as it is also no longer `location` aware. This is a temporary
hack, not a feature ;)