Async fetch of equipment status & sync time #8 (#23)
* Sync current time every 10s

Would have preferred to use setInterval in the store's action directly,
but encountered a bug where the action was unsuccessful in committing a
mutation called within setInterval. This is not reproducible in vanilla
Vue, and appears to be isolated to Nuxt. I have some demos in glitch
which I will link in the PR.

Add styles borrowed from Mann's circdisplay for blinking colon.

Also, hat tip to Jonnie Hallman for the great post on reactive time with
Vue.js [1].

[1] https://cushionapp.com/journal/reactive-time-with-vuejs

* Update equipment status every minute

Again, would have preferred to use setInterval within the action of the
store vs the component. See 3b353f8 for more details.

* Syntax tweaks for element closure & css animation

In response to Manolo's review of #23.