Mann consultation screens for tablets (#34)
* First pass at Mann consultation screens

Bare minimum. Hit LibCal API and store response in Vuex for single desk.
Temporarily using API's `currently_open` field to determine status,
which we know is a big tease [1]...if only it were that simple.

BG color change based on status via Nuxt's header method, which stands
on the shoulders of `vue-meta`. Use fiord (`#485963`) from original
mockups for `by-appointment` status, which we hadn't considered in early
phases, but GIS holds this status the entire summer. No styling beyond
this and Lato font.

[1] cul-it/circ-display@53e6f4e

* Don't rely on LibCal's `currently_open` status

Since we'll be caching, determine current status manually.

* checks for by appointment status

* moves axios fetch out of Vuex store and into util

* prep for querying LibCal to retrieve hours other than today

* Return current status & change via nextOpen()

Closing time can be returned as the status.change thanks to switching
from `.findIndex()` to `.find()`. Next up is to determine the opening
time for any closed or by appt desks.

* Next opening time for any closed desk

Check the next two weeks (14 days).

* Extract timeFormat to default export

So it can be used by all... (DRY)

* Placeholder text when no upcoming openings

* Include today in nextOpening & momentjs formatting

* Inject desk name as page title via head (vue-meta)

* First pass at styles based on mockups

No layout just yet. Will be using CSS grid. Also letter-spacing will
need to be adjust per desk. For now, using value based on testing with
CISER.

* CSS Grid FTW

* Consistent display format for time

* If dealing with today, ensure not already closed

TODO: Review immutable.js [1] and seamless-immutable [2]. For now,
workaround is to copy hours array before using it as a parameter for
alreadyClosed(), which results in clearing out the original array.

[1] https://facebook.github.io/immutable-js
[2] https://github.com/rtfeldman/seamless-immutable

* Letter-spacing per desk